pyproject.toml
#

[tool.black]
line-length = 79

[tool.poetry]
name = "stonkers"
version = "0.0.1"
description = "Stonkers"
authors = ["Jason Kölker <jason@koelker.net>"]
license = "MIT"
readme = "README.rst"
homepage = "https://github.com/jkoelker/stonkers"
repository = "https://github.com/jkoelker/stonkers"
classifiers = [
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.urls]
Changelog = "https://github.com/jkoelker/stonkers/releases"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
tda-api = "^1.6.0"
pyyaml = "^6.0.0"
pandas = "^2.1.4"
tabulate = "^0.9.0"
scipy = "^1.11.4"
click = "^8.1.3"
httpx = "^0.26.0"
rich = "^13.7.0"
async-lru = "^2.0.4"
cachetools = "^5.3.2"

[tool.poetry.scripts]
stonkers = "stonkers.cli:cli"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
black = "^23.12.1"
ipython = "^8.19.0"
datascroller = "^1.4.1"
mypy = "^1.8.0"
types-pyyaml = "^6.0.12.12"
pandas-stubs = "^2.1.4.231227"
types-cachetools = "^5.3.0.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "Poetry.core.masonry.api"

[tool.pylint."messages control"]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-many-arguments",
]
